Updated flinks as a separate element
